do not use global instance of chatstore in EmptyChatMessage, also move it to own file
