+++ src/renderer/components/message/MessageList.tsx
@@ -25,6 +25,7 @@ import { MessagesDisplayContext } from '../../contexts/MessagesDisplayContext'
 import useTranslationFunction from '../../hooks/useTranslationFunction'
 import useKeyBindingAction from '../../hooks/useKeyBindingAction'
 import { useReactionsBar } from '../ReactionsBar'
+import EmptyChatMessage from './EmptyChatMessage'
 
 const log = getLogger('render/components/message/MessageList')
 
@@ -528,7 +529,9 @@ export const MessageListInner = React.memo(
         onScroll={onScroll}
       >
         <ul>
-          {messageListItems.length === 0 && <EmptyChatMessage />}
+          {messageListItems.length === 0 && (
+            <EmptyChatMessage chatStore={chatStore} />
+          )}
           {activeView.map(messageId => {
             if (messageId.kind === 'dayMarker') {
               return (
@@ -646,40 +649,6 @@ function JumpDownButton({
         </div>
       </div>
     </>
-  )
-}
-
-function EmptyChatMessage() {
-  const tx = useTranslationFunction()
-  const chatStore = useChatStore()
-  const chat = chatStore.chat
-
-  if (!chat) {
-    throw new Error('no chat selected')
-  }
-
-  let emptyChatMessage = tx('chat_new_one_to_one_hint', [chat.name, chat.name])
-
-  if (chat.chatType === C.DC_CHAT_TYPE_BROADCAST) {
-    emptyChatMessage = tx('chat_new_broadcast_hint')
-  } else if (chat.chatType === C.DC_CHAT_TYPE_GROUP && !chat.isContactRequest) {
-    emptyChatMessage = chat.isUnpromoted
-      ? tx('chat_new_group_hint')
-      : tx('chat_no_messages')
-  } else if (chat.isSelfTalk) {
-    emptyChatMessage = tx('saved_messages_explain')
-  } else if (chat.isDeviceChat) {
-    emptyChatMessage = tx('device_talk_explain')
-  } else if (chat.isContactRequest) {
-    emptyChatMessage = tx('chat_no_messages')
-  }
-
-  return (
-    <li>
-      <div className='info-message big'>
-        <div className='bubble'>{emptyChatMessage}</div>
-      </div>
-    </li>
   )
 }
